Extract admin-lens Firestore ops into host/admin-firestore.ts
Move admin.getDocument/listDocuments/setDocument/deleteDocument/readState into
host/admin-firestore.ts behind isAdminFirestoreOp/handleAdminFirestoreOp.
handleOp routes to it. Bodies verbatim; no behavior change.
